NAS-123271: Move interfaces from webworker
