+++ src/app/interfaces/reporting.interface.ts
@@ -139,3 +139,14 @@ export interface ReportingData {
     [key in ReportingAggregationKeys]: ReportingAggregationValue;
   };
 }
+
+export interface Command {
+  command: string; // Use '|' or '--pipe' to use the output of previous command as input
+  input: unknown;
+  options?: unknown[]; // Function parameters
+}
+
+export enum ReportingDatabaseError {
+  FailedExport = 22,
+  InvalidTimestamp = 206,
+}
